+++ Sources/QuillKit/Operations/Operation.swift
@@ -1,0 +1,7 @@
+import Foundation
+
+public protocol Operation {
+    associatedtype Body
+    associatedtype Response
+    var route: Router { get }
+}
